+++ CHANGELOG.md
@@ -21,6 +21,16 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v1.1.4
+
+Released Friday, 17th April 2020.
+
+### Dependencies
+
+* Upgraded to '@ganbarodigital/ts-lib-error-reporting' v0.3
+* Upgraded to '@ganbarodigital/ts-lib-packagename' v0.3
+* Upgraded to '@ganbarodigital/ts-lib-http-types' v0.3
+
 ## v1.1.3
 
 Released Monday, 6th April 2020.